pipelines/meta-components/aixcc-sc-capi/pyproject.toml
[tool.poetry]
name = "competition-api"
version = "0.1.0"
description = ""
authors = ["Thing 012 <thing-012@aicyberchallenge.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
prestart = "competition_api.prestart:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.12"

aiofile = "^3.8.8"
alembic = "^1.13.1"
argon2-cffi = "^23.1.0"
asyncpg = "^0.29.0"
fastapi = "^0.111.0"
gitpython = "^3.1.43"
psycopg2-binary = "^2.9.9"
pydantic = "^2.7.1"
ruamel-yaml = "^0.18.6"
sqlalchemy = {version = "^2.0.30", extras = ["asyncio"]}
sqlalchemy-dlock = "^0.6.1.post1"
structlog = "^24.1.0"
types-whatthepatch = "^1.0.2.5"
uvicorn = "^0.30.0"
uvloop = "^0.19.0"
vyper-config = "^1.2.1"
whatthepatch = "^1.0.5"

[tool.poetry.group.dev.dependencies]
bandit = ">=1.7.8"
black = ">=24.4.2"
isort = ">=5.13.2"
mypy = ">=1.10.0"
pylint = ">=3.1.0"
pytest = ">=8.2.0"
pytest-asyncio = "<0.23" # https://github.com/pytest-dev/pytest-asyncio/issues/706
pytest-cov = ">=5.0.0"
pytest-docker-tools = ">=3.1.3"
safety = ">=3.2.0"
types-psycopg2 = ">=2.9.21.20240417"

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["competition_api","tests"]

[tool.black]
line-length = 88

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-branches",
    "too-many-locals",
    "too-many-statements",
    "import-error", # superlinter does not run in our venv
    "fixme",
    "invalid-name"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
allow_redefinition = true
check_untyped_defs = true
ignore_missing_imports = true # superlinter does not run in our venv
